return types, 64-bit changes: ioctl cmd -> u_long, casts, etc.
SCCS-vsn: sys/kern/subr_log.c 8.2
SCCS-vsn: sys/kern/tty_tb.c 8.2
SCCS-vsn: sys/kern/tty.c 8.13
SCCS-vsn: sys/kern/kern_sig.c 8.12
SCCS-vsn: sys/kern/tty_pty.c 8.3
SCCS-vsn: sys/kern/kern_resource.c 8.7
SCCS-vsn: sys/kern/sys_socket.c 8.2
SCCS-vsn: sys/kern/vfs_bio.c 8.11
SCCS-vsn: sys/kern/uipc_usrreq.c 8.6
SCCS-vsn: sys/kern/kern_prot.c 8.8
SCCS-vsn: sys/kern/uipc_mbuf.c 8.3
SCCS-vsn: sys/kern/kern_proc.c 8.6
SCCS-vsn: sys/kern/tty_tty.c 8.3
SCCS-vsn: sys/kern/tty_compat.c 8.2
SCCS-vsn: sys/kern/vfs_subr.c 8.14
SCCS-vsn: sys/kern/tty_conf.c 8.5
SCCS-vsn: sys/kern/vfs_vnops.c 8.6

+++ usr/src/sys/kern/sys_socket.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)sys_socket.c	8.1 (Berkeley) %G%
+ *	@(#)sys_socket.c	8.2 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -48,7 +48,7 @@ soo_write(fp, uio, cred)
 
 soo_ioctl(fp, cmd, data, p)
 	struct file *fp;
-	int cmd;
+	u_long cmd;
 	register caddr_t data;
 	struct proc *p;
 {
